Feature: PAX support visibility map for delete/update
PAX will rewrite the whole data file if any tuple in the data
file will be deleted/updated. The benefit is that all tuples
in the data file are visible, the column values are continuous.
The side effect amplifies write overhead and will slow scanning
tuples a bit.

The major bad side is that rewriting the whole data file will
bloat the index file. Using visibility map could reduce
the overhead for write and be friendly for index.

+++ contrib/pax_storage/src/cpp/access/pax_deleter.cc
@@ -5,19 +5,24 @@
 #include <vector>
 
 #include "access/pax_dml_state.h"
+#include "access/paxc_rel_options.h"
 #include "catalog/pax_aux_table.h"
 #include "comm/singleton.h"
 #include "storage/pax_itemptr.h"
 namespace pax {
 CPaxDeleter::CPaxDeleter(Relation rel, Snapshot snapshot)
-    : rel_(rel), snapshot_(snapshot) {}
+    : rel_(rel), snapshot_(snapshot) {
+  delete_xid_ = GetCurrentTransactionId();
+  Assert(TransactionIdIsValid(delete_xid_));
+
+  use_visimap_ = true;
+}
 
 TM_Result CPaxDeleter::DeleteTuple(Relation relation, ItemPointer tid,
                                    CommandId cid, Snapshot snapshot,
                                    TM_FailureData *tmfd) {
   CPaxDeleter *deleter =
       CPaxDmlStateLocal::Instance()->GetDeleter(relation, snapshot);
-  // TODO(gongxun): need more graceful way to pass snapshot
   Assert(deleter != nullptr);
   TM_Result result;
   result = deleter->MarkDelete(tid);
@@ -26,6 +31,7 @@ TM_Result CPaxDeleter::DeleteTuple(Relation relation, ItemPointer tid,
   }
   return result;
 }
+
 // used for delete tuples
 TM_Result CPaxDeleter::MarkDelete(ItemPointer tid) {
   uint32 tuple_offset = pax::GetTupleOffset(*tid);
@@ -34,9 +40,11 @@ TM_Result CPaxDeleter::MarkDelete(ItemPointer tid) {
 
   if (block_bitmap_map_.find(block_id) == block_bitmap_map_.end()) {
     block_bitmap_map_[block_id] =
-        pax_shared_ptr<Bitmap64>(PAX_NEW<Bitmap64>());  // NOLINT
-    cbdb::DeleteMicroPartitionEntry(RelationGetRelid(rel_), snapshot_,
+        pax_shared_ptr<Bitmap8>(PAX_NEW<Bitmap8>());  // NOLINT
+    if(!use_visimap_) {
+      cbdb::DeleteMicroPartitionEntry(RelationGetRelid(rel_), snapshot_,
                                     block_id);
+    }
   }
   auto bitmap = block_bitmap_map_[block_id].get();
   if (bitmap->Test(tuple_offset)) {
@@ -46,12 +54,23 @@ TM_Result CPaxDeleter::MarkDelete(ItemPointer tid) {
   return TM_Ok;
 }
 
+bool CPaxDeleter::IsMarked(ItemPointerData tid) const {
+  std::string block_id = MapToBlockNumber(rel_, tid);
+  auto it = block_bitmap_map_.find(block_id);
+
+  if (it == block_bitmap_map_.end()) return false;
+
+  const auto &bitmap = it->second;
+  uint32 tuple_offset = pax::GetTupleOffset(tid);
+  return bitmap->Test(tuple_offset);
+}
+
 // used for merge remaining partition files, no tuple needs to delete
 void CPaxDeleter::MarkDelete(BlockNumber pax_block_id) {
   std::string block_id = std::to_string(pax_block_id);
 
   if (block_bitmap_map_.find(block_id) == block_bitmap_map_.end()) {
-    block_bitmap_map_[block_id] = pax_shared_ptr<Bitmap64>(PAX_NEW<Bitmap64>());
+    block_bitmap_map_[block_id] = pax_shared_ptr<Bitmap8>(PAX_NEW<Bitmap8>());
     cbdb::DeleteMicroPartitionEntry(RelationGetRelid(rel_), snapshot_,
                                     block_id);
   }
@@ -62,7 +81,11 @@ void CPaxDeleter::ExecDelete() {
 
   TableDeleter table_deleter(rel_, BuildDeleteIterator(), block_bitmap_map_,
                              snapshot_);
-  table_deleter.Delete();
+  if (use_visimap_) {
+    table_deleter.DeleteWithVisibilityMap(delete_xid_);
+  } else {
+    table_deleter.Delete();
+  }
 }
 
 pax_unique_ptr<IteratorBase<MicroPartitionMetadata>>
